Tarea del 18/10/2022 con float para ci

+++ AplicacionArbol1810.java
@@ -9,14 +9,14 @@ public class AplicacionArbol1810 {
     public static void main(String[] args) {
 
         Arbol1810 arbol = new Arbol1810();
-        int valor;
+        float valor;
         System.out.println("Insertando los siguientes valores");
         //insertar 10 enteros aleatorios del 0 al 99 en arbol
 
         for (int i = 1; i <= 10; i++) {
-            valor = (int) (Math.random() * 100);
+            valor = (float) (Math.random() * 100);
             System.out.println(valor + " ");
-            arbol.insertarNodo(valor);
+            arbol.insertarNodo((int) valor);
         }
         System.out.println("\n\nRecorrido Preorden");
         arbol.recorridoPreorden();//realizar el recorrido preorden de arbol
